oct-2-超声乳腺图像的BIRADS分类&特征识别/requirements.txt
matplotlib==3.8.3
numpy==1.26.4
Pillow==10.3.0
scikit-learn==1.5.0
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.0
tqdm==4.66.2
opencv-python
pandas